\name{resourceVersions}
\alias{resourceVersions}
\title{
    Version numbers of DGIdb resources
}
\description{
    Prints the version numbers of all DGIdb resources.
}
\usage{
    resourceVersions()
}
\value{
    Returns a \code{data.frame} with two columns, the name and the 
    version numbers of the resources.
}
\author{
    Thomas Thurnherr <thomas.thurnherr@gmail.com>
}
\seealso{
    \url{http://dgidb.genome.wustl.edu/sources}
}
\examples{
    resourceVersions()
}
\keyword{ ~data }
\keyword{ ~print }
\keyword{ ~documentation }